+++ src/modules/footer/components/Footer.vue
@@ -20,6 +20,7 @@ export default {
     },
     computed: {
         ...mapGetters(["footerConfig", "mobile", "masterPortalVersionNumber"]),
+        ...mapGetters("Map", ["is3d"]),
         ...mapGetters("Footer", Object.keys(getters)),
         showLanguageSwitcher () {
             return this.$i18n.i18next.options.isEnabled() && Object.keys(this.$i18n.i18next.options.getLanguages()).length > 1;
@@ -62,6 +63,17 @@ export default {
                     model.setIsActive(!model.attributes.isActive);
                 }
             }
+        },
+        supportedIn3D (toolModelId) {
+            if (!toolModelId) {
+                return true;
+            }
+            if (this.is3d) {
+                const toolsSupportedIn3d = Radio.request("Tool", "getSupportedIn3d");
+
+                return toolsSupportedIn3d.find(name => name.toLowerCase() === toolModelId.toLowerCase());
+            }
+            return true;
         }
     }
 };
@@ -77,6 +89,7 @@ export default {
         <template v-if="showFooter">
             <template v-for="(url, index) in urls">
                 <span
+                    v-if="supportedIn3D(url.toolModelId)"
                     :key="`footer-url-${index}`"
                 >
                     {{ $t(url.bezeichnung) }}